+++ noodles-vcf/src/header/record/value/map/other.rs
@@ -1,7 +1,5 @@
 use std::fmt;
 
-use indexmap::IndexMap;
-
 use super::{builder, tag, Fields, Inner, Map, TryFromFieldsError};
 
 type StandardTag = tag::Identity;
@@ -9,9 +7,7 @@ type Tag = tag::Tag<StandardTag>;
 
 /// An inner VCF header other map value.
 #[derive(Clone, Debug, Default, Eq, PartialEq)]
-pub struct Other {
-    id: String,
-}
+pub struct Other;
 
 impl Inner for Other {
     type StandardTag = StandardTag;
@@ -25,16 +21,10 @@ impl Map<Other> {
     ///
     /// ```
     /// use noodles_vcf::header::record::value::{map::Other, Map};
-    /// let map = Map::<Other>::new("noodles");
+    /// let map = Map::<Other>::new();
     /// ```
-    pub fn new<I>(id: I) -> Self
-    where
-        I: Into<String>,
-    {
-        Self {
-            inner: Other { id: id.into() },
-            other_fields: IndexMap::new(),
-        }
+    pub fn new() -> Self {
+        Self::default()
     }
 }
 
@@ -50,19 +40,15 @@ impl TryFrom<Fields> for Map<Other> {
     fn try_from(fields: Fields) -> Result<Self, Self::Error> {
         let mut other_fields = super::init_other_fields();
 
-        let mut id = None;
-
         for (key, value) in fields {
             match Tag::from(key) {
-                Tag::Standard(StandardTag::Id) => super::parse_id(&value, &mut id)?,
+                Tag::Standard(StandardTag::Id) => return Err(TryFromFieldsError::DuplicateTag),
                 Tag::Other(t) => super::insert_other_field(&mut other_fields, t, value)?,
             }
         }
 
-        let id: String = id.ok_or(TryFromFieldsError::MissingField("ID"))?;
-
         Ok(Self {
-            inner: Other { id },
+            inner: Other,
             other_fields,
         })
     }
@@ -74,23 +60,24 @@ mod tests {
 
     #[test]
     fn test_fmt() {
-        let map = Map::<Other>::new("noodles");
+        let map = Map::<Other>::new();
         assert!(map.to_string().is_empty());
     }
 
     #[test]
-    fn test_try_from_fields_for_map_other() -> Result<(), TryFromFieldsError> {
-        let actual = Map::<Other>::try_from(vec![(String::from("ID"), String::from("noodles"))])?;
-        let expected = Map::<Other>::new("noodles");
+    fn test_try_from_fields_for_map_other() -> Result<(), Box<dyn std::error::Error>> {
+        let actual = Map::<Other>::try_from(vec![(String::from("noodles"), String::from("vcf"))])?;
+
+        let noodles_tag = match Tag::from(String::from("noodles")) {
+            Tag::Other(tag) => tag,
+            Tag::Standard(_) => panic!("invalid tag"),
+        };
+
+        let expected = Map::<Other>::builder()
+            .insert(noodles_tag, String::from("vcf"))
+            .build()?;
+
         assert_eq!(actual, expected);
         Ok(())
-    }
-
-    #[test]
-    fn test_try_from_fields_for_info_with_missing_fields() {
-        assert_eq!(
-            Map::<Other>::try_from(Vec::new()),
-            Err(TryFromFieldsError::MissingField("ID"))
-        );
     }
 }