+++ noodles-vcf/CHANGELOG.md
@@ -13,6 +13,10 @@
 
   * vcf/header/info: Move `Key` under `record::info::field` module.
 
+  * vcf/header/record/value/map/other: Move record ID from record to map key.
+
+    `Map<Other>` no longer holds the record ID. Use the collection key instead.
+
   * vcf/indexed_reader/builder: Attempt to load either an associated tabix
     index (`<src>.tbi`) or CSI (`<src>.csi`), in that order.